Squashed bugs.  Player now looks in the correct direction, when time reverses so does all the clones of MC (problem was that I was using an unsigned type and setting it to -1... so it would explode to 2 billion something... damn size_t's!!! They do that to me EVERY TIME!

+++ sprite.h
@@ -35,7 +35,7 @@ protected:
 
     bool m_useSlice;
 
-    size_t m_sliceBegin, m_sliceEnd, m_sliceIndex;
+    int m_sliceBegin, m_sliceEnd, m_sliceIndex;
 
     void beginSlice() {
         // Set the first state we want to be the last state in the stack
